Add devDependency @types/body-parser
This commit adds the development dependency @types/body-parser to the project. The package.json file has been updated to include this dependency, and a corresponding entry has been added to the package-lock.json file. Additionally, the devDependency has been saved in a new "devDependencies" object within the package-lock.json file.