Update User model with email, password, and role properties; define and export User model
The User model interface is updated to include email, password, and role properties. The email property is set to be unique and required. The password property is set to be required and have a minimum length of 6 characters. The role property is set to a default value of 'user' and can only be 'user' or 'admin'. The Admin model is removed and replaced with the updated User model.